Update risk-management.rst
'Move CW20 online  - We would have to ask sponsors' - reword e.g. seek sponsors' approval

'Ultimately, further Government restrictions and lockdown measures throughout March would have prevented the event taking place in person if we had decided to go ahead with an in-person event.' - this is a repeat from the venue section, remove one instance

'We note that we did not change our privacy policy or apply a license to our collaborative documents, which required us to "sanitise" them before sharing in compliance with GDPR.' - are you saying you broke GDPR? If so don't say this publicly! If not then rephrase to clarify.